refactor: restructure project into bounded contexts with independent crates
- Introduced a new architecture by organizing the project into bounded contexts, each implemented as a separate Rust crate.
- Created `library` crate for the library management context, encapsulating domain, application, infrastructure, and presentation layers.
- Established a `shared` crate for common utilities and error handling across contexts.
- Updated the main application (`lifebook`) to integrate these changes, including adjustments to dependency management and file structure.
- Removed legacy module structure to streamline the codebase and improve maintainability.
