+++ src-tauri/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["lifebook", "migration", "entity"]
+members = ["lifebook", "migration", "entity", "contexts/library", "contexts/shared"]
 resolver = "2"
