+++ src-tauri/contexts/library/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "library"
+version = "0.1.0"
+edition = "2024"
+publish = false
+
+[lib]
+name = "library"
+path = "src/lib.rs"
+
+[dependencies]
+shared = { path = "../shared" }
+entity = { path = "../../entity" }
+async-graphql = "7"
+async-trait = "0.1"
+chrono = "0.4"
+sea-orm = { version = "1", features = ["sqlx-sqlite", "runtime-tokio-rustls", "macros"] }
+
