+++ src-tauri/contexts/library/src/application/dto.rs
@@ -1,0 +1,3 @@
+// Library Application Layer - DTOs
+
+pub mod book;
