+++ src-tauri/contexts/library/src/application/services.rs
@@ -1,0 +1,3 @@
+// Library Application Layer - Services
+
+pub mod book;
