+++ src-tauri/contexts/library/src/domain/entities.rs
