+++ src-tauri/contexts/library/src/infrastructure/repositories.rs
@@ -1,0 +1,3 @@
+// Library Infrastructure Layer - Repositories
+
+pub mod book;
