+++ src-tauri/contexts/library/src/presentation/presentation.rs
@@ -1,0 +1,4 @@
+// Library Presentation Layer
+
+#[path = "graphql/graphql.rs"]
+pub mod graphql;
