+++ src-tauri/contexts/shared/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "shared"
+version = "0.1.0"
+edition = "2024"
+publish = false
+
+[lib]
+name = "shared"
+path = "src/lib.rs"
+
+[dependencies]
+
