+++ src-tauri/contexts/shared/src/application/application.rs
