+++ src-tauri/contexts/shared/src/domain/domain.rs
