+++ src-tauri/contexts/shared/src/domain/errors.rs
