+++ src-tauri/contexts/shared/src/lib.rs
@@ -1,0 +1,7 @@
+// Shared Context - 全コンテキスト共通
+
+#[path = "application/application.rs"]
+pub mod application;
+
+#[path = "domain/domain.rs"]
+pub mod domain;
