--- src-tauri/lifebook/src/modules.rs
@@ -1,4 +1,0 @@
-// Modules - ドメインモジュール
-
-pub mod library;
-pub mod shared;
