--- src-tauri/lifebook/src/modules/library.rs
@@ -1,5 +1,0 @@
-// Library Module - 図書管理の境界づけられたコンテキスト
-
-pub mod application;
-pub mod domain;
-pub mod infrastructure;
