--- src-tauri/lifebook/src/modules/library/application/dto.rs
@@ -1,3 +1,0 @@
-// Library Application - DTOs
-
-pub mod book;
