--- src-tauri/lifebook/src/modules/library/application/services.rs
@@ -1,3 +1,0 @@
-// Library Application - Services
-
-pub mod book;
