--- src-tauri/lifebook/src/modules/library/infrastructure/repositories.rs
@@ -1,3 +1,0 @@
-// Library Infrastructure - Repositories
-
-pub mod book;
