--- src-tauri/lifebook/src/modules/shared.rs
@@ -1,4 +1,0 @@
-// Shared Module - 全コンテキスト共通
-
-pub mod application;
-pub mod domain;
