--- src-tauri/lifebook/src/presentation.rs
@@ -1,4 +1,0 @@
-// Presentation Layer
-
-pub mod library;
-pub mod schema;
