--- src-tauri/lifebook/src/presentation/library.rs
@@ -1,4 +1,0 @@
-// Presentation Layer - Library Module
-
-pub mod mutations;
-pub mod queries;
